NPC and trading menu started
player can interact with an npc which brings up a trading menu. player can click an item in their hotbar, which moves it to the trading menu, or click the trading menu slot, which moves the held item back to th eplayer's inventory. closing the trading menu moves the item back to the player's inventory.